feat: Add user authentication and logout functionality
Refactor the Navbar component to include user authentication and logout functionality. This includes adding the ability to display different navigation options based on whether the user is logged in or not. The Logout component is also added to handle the logout process. Additionally, the authSlice.js file is updated to include the clearUser action for clearing the user state. The Logout component is added to the Navbar component to allow users to log out of the application.